Dataretrieval fixed
The retrieval of data from mongoDB should place buyers city in pairs with the sellers that they've purchased stuff from.
Meaning if buyer0 has bought items from seller0 and seller1 the array will be buyersandseller[0][0] = ['buyer0', 'seller0'] and so on. The allWaypoints array is sorted the same way so the route in the grid is sorted with the same numbers as the corresponding pair of buyer/seller